+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Add `an={account}` query string to VTEX ID authentication and user info calls.
+
 ## [2.177.1] - 2026-05-11
 
 ### Added